requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
jinja2>=3.1.0
requests>=2.31.0
urllib3>=2.0.0